market-research-platform/backend/requirements.txt
# filepath: market-research-platform/backend/requirements.txt
# Pin versions for reproducible builds.

# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── LlamaIndex core ─────────────────────────────────────────────────────────
llama-index-core==0.10.40
llama-index-llms-openai==0.1.22
llama-index-embeddings-openai==0.1.10
llama-index-vector-stores-postgres==0.1.6
llama-index-readers-web==0.1.19

# ─── LlamaParse (PDF parsing — optional, cloud API) ─────────────────────────
llama-parse==0.4.7

# ─── Fallback PDF parsing (local) ───────────────────────────────────────────
pdfplumber==0.11.0
PyPDF2==3.0.1

# ─── OpenAI client ───────────────────────────────────────────────────────────
openai==1.30.0

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5

# ─── News ingestion ───────────────────────────────────────────────────────────
newsapi-python==0.2.7
feedparser==6.0.11

# ─── Task scheduling ──────────────────────────────────────────────────────────
apscheduler==3.10.4

# ─── Report export ────────────────────────────────────────────────────────────
reportlab==4.2.0
python-docx==1.1.2

# ─── Utilities ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
structlog==24.1.0
